rl_mlpbook/requirements.txt
gym==0.14.0
gym_ple
jupyter==1.0.0
numpy==1.16.4
pandas==0.24.2
scipy==1.3.0
scikit-learn==0.21.2
matplotlib==3.1.0
#  tensorflow==1.14.0
tensorflow-gpu==1.14.0
-e git+https://github.com/ntasfi/PyGame-Learning-Environment.git#egg=ple
-e git+https://github.com/lusob/gym-ple.git#egg=gym-ple
h5py==2.9.0
pygame==1.9.6
tqdm==4.32.1
torch==1.3.0
pycodestyle
